Global rewrite (again), now back in Python.
Discord Alchemey is dead, and the most recent version doesn0t work with
the latest version of Elixir. As there appears to be no other Elixir
package for working with Discord and I don’t have the patience (or
expertise) to maintain Discord Alchemy, a rewrite is needed.

The rewritten version is much more concise, much easier to read, fixes a
few bugs in the previous code, and also adds a few new QoL features.
It’s currently mostly feature complete relative to the old code, though
may have new bugs I didn’t notice so far in testing.