Increase reroll limit from 8 to 32.
Individual reroll attempts are in general not exceptionally expensive,
but having a low limit increases the probability of failing to find a
valid item within the requested price range when asking for items based
on cost limits.

Increasing the limit should have little to no impact on the common case,
but should enable much narrower cost limits to work more frequently.